requirements.txt
attrs>=19.1.0,<19.99.99
ciso8601>=2.1.1,<2.99.99
cryptography>=2.4.1,<3
Django>=2.1.5,<3
django-cors-headers>=2.4.0,<2.99.99
psycopg2-binary==2.7.6
PyJWT>=1.6.4,<2
gunicorn>=19.9.0,<19.99.99
twilio>=6.26.3,<6.99.99
phonenumberslite>=8.10.11,<8.99.99
